+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+## [2.12.3] - 2021-08-05
+
 ### Added
 
 - Option to hide Pagination "Posts Per Page" text in app settings.